Fix: bugs

+++ train.py
@@ -112,7 +112,7 @@ def train(args):
             ###########################
             optimizer_D.zero_grad()
 
-            label = label.squeeze(1)
+            label = label.squeeze(1).float()
 
             # train with real
             real_label = netD(real_S)
